pyproject.toml
[tool.poetry]
name = "app-name"
version = "0.1.0"
description = "sample devcontainer repo"
authors = ["Dipsy"]
readme = "README.md"
packages = [{ include = "server", from = "src" }]

[tool.poetry.dependencies]
python="^3.12"
fastapi = "0.110.1"
uvicorn = "^0.29.0"
httpx = "^0.27.0"
kafka-python = "^2.0.2"
six = "^1.16.0"
kafka-python-ng = "^2.2.3"
clickhouse-driver = "^0.2.9"
clickhouse-connect = "^0.8.7"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.4.4"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "server.main:start"

[tool.black]
line-length = 100
target-version = ["py312"]
